requirements.txt
Django==4.2.3
redis==4.6.0
celery==5.3.1
flower==2.0.0
psycopg2-binary==2.9.6
Pillow==10.0.0

social-auth-app-django==5.2.0
django-cors-headers==4.2.0
django-filter==23.2
django-environ==0.10.0
django-silk==5.0.3
django-phonenumber-field[phonenumberslite]==7.1.0
django-autoslug==1.9.9
django-ckeditor==6.7.0

djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.4
drf-extra-fields==3.6.1

flake8-django==1.3
Flake8-pyproject==1.2.3
isort==5.12.0
black==23.7.0

Sphinx==7.1.2
sphinx-rtd-theme==1.3.0